updting pointer

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: csantos- <csantos-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/12 00:27:07 by csantos-          #+#    #+#             */
-/*   Updated: 2021/04/12 23:17:05 by csantos-         ###   ########.fr       */
+/*   Updated: 2021/04/13 20:49:31 by csantos-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,7 +56,7 @@ void			print_choi(t_flags *flags, int c);
 void			print_padding(t_flags *flags, int len);
 void			print_s(t_flags *flags, char *s);
 void			print_doido_da_nat(t_flags *flags, char *number, int size);
-void			print_d_i(t_flags *flags, int num);
+void			print_d_i(t_flags *flags, long int num);
 void			print_du_luigi(t_flags *flags, long int num);
 void			print_hex(t_flags *flags, long int num);
 void			print_doido_da_cla(t_flags *flags, char *number, int size);
